Back port changes from: 191fd21
